--- algofi-compile/src/main/java/gooroommoon/algofi_compile/api/CheckoutApiController.java
@@ -1,4 +1,0 @@
-package gooroommoon.algofi_compile.api;
-
-public class CheckoutApiController {
-}
